audio component 분리리
